+++ riptide/tests/run_tests.py
@@ -6,6 +6,6 @@ def test():
     """
     Run the test suite.
     """
-    tests_dir = os.path.join(os.path.dirname(__file__), '..', 'tests')
+    tests_dir = os.path.dirname(__file__)
     args = ['--verbose', tests_dir]
     pytest.main(args)